requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.0
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.3.1
